feat(x/gov): validators can only vote with own stake (#20)
Change #7 so that validators are allowed to vote, but cannot use
delegations in governance to increase their voting power, they can only
vote with their own stake. This seems more sensible wrt the previous
solution of removing their ability to vote entirely.

This also addresses #19 as I added comments in the README, and also che
changelog should be good. Therefore with this PR we should be able to
also merge the feat branch back into main and close #6