test: improve vote inheritance testing (#23)
While reviewing #20, I found that the tally result assertion was weak,
since it only ensures that the expected tally result is not empty.

This change improves that for tests that imply vote inheritance, the
first commit e257af4396268284d0b382103ce754b481d2fe12 re-enables vote
inheritance and improves the tally result assertion. The second commit
3696c7cf6c1f7178ad365af0fc9075db4cbc8b34 disables vote inheritance again
and updates the assertions to show that vote inheritance is correctly
disabled (removing the `delTokens` additions in the expected tally
results).